Модуль 13
- Логика таймера вынесена в отдельный модуль timer
