+++ index.html
@@ -43,7 +43,7 @@
         <button class="btn btn-reset hide" id="btn-reset">Сбросить</button>
     </div>
 
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 
 </html>
